+++ playground/src/main.js
@@ -4,4 +4,5 @@ import './style.css'
 // 导入 @vue/repl 的样式
 import '@vue/repl/style.css'
 
+
 createApp(App).mount('#app')